Use NeuronCore's entity id instead of declaring a second one
Neuron::INVALID_ENTITY_ID is NeuronCore's, a u16 beside EntityRecord::id.
SceneEntity declared its own as a u32, and every translation unit that saw both
failed with C2371. The two files have different names, so ADR-013 §3's
uniqueness check passed it -- but namespace-scope constants share one flat
namespace exactly the way file names share one flat include path, and for the
same reason.

Sixteen bits is the right width anyway: an id is EntityRecord::id, and
OrderIntent::entityIds is a const uint16_t*, so a selection now feeds an order
with no conversion at all -- which is what S9 needs of it next.

CI gained the check the file rule could not make: two engine headers declaring
the same namespace-scope constant now fail the build, with both locations
named. Class members are exempt, being scoped by their class. Fourteen
constants today, all declared once. ADR-013 §3b records it.

My harness missed this because it did not include NeuronCore's real headers
alongside NeuronClient's; it does now, so the next collision of this shape
fails locally instead.

+++ NeuronClient/OverlayMark.cpp
@@ -14,7 +14,7 @@ namespace
 /// The entity with this id, or null. Linear because a selection is a handful
 /// of ships and the entity list is a few hundred: a map would cost more to
 /// build every frame than the scan costs to run.
-[[nodiscard]] const SceneEntity* FindEntity(std::span<const SceneEntity> _entities, std::uint32_t _id) noexcept
+[[nodiscard]] const SceneEntity* FindEntity(std::span<const SceneEntity> _entities, std::uint16_t _id) noexcept
 {
   const auto found = std::find_if(_entities.begin(), _entities.end(), [_id](const SceneEntity& _e) { return _e.id == _id; });
   return found == _entities.end() ? nullptr : &*found;
@@ -30,7 +30,7 @@ namespace
 
 } // namespace
 
-void BuildOverlayMarks(std::span<const SceneEntity> _entities, std::span<const std::uint32_t> _selectedIds,
+void BuildOverlayMarks(std::span<const SceneEntity> _entities, std::span<const std::uint16_t> _selectedIds,
                        const OverlayTuning& _tuning, float _metresPerPixel, OverlayMarkList& _outMarks)
 {
   _outMarks.Clear();
@@ -43,7 +43,7 @@ void BuildOverlayMarks(std::span<const SceneEntity> _entities, std::span<const s
   // different depth state and the split has to be a contiguous range.
   const float minRadiusMetres = _tuning.ringMinRadiusPixels * _metresPerPixel;
 
-  for (const std::uint32_t id : _selectedIds)
+  for (const std::uint16_t id : _selectedIds)
   {
     const SceneEntity* entity = FindEntity(_entities, id);
     if (entity == nullptr)
@@ -60,7 +60,7 @@ void BuildOverlayMarks(std::span<const SceneEntity> _entities, std::span<const s
   }
   _outMarks.ringCount = static_cast<std::uint32_t>(_outMarks.marks.size());
 
-  for (const std::uint32_t id : _selectedIds)
+  for (const std::uint16_t id : _selectedIds)
   {
     const SceneEntity* entity = FindEntity(_entities, id);
     if (entity == nullptr)
