+++ NeuronClient/Selection.cpp
@@ -25,17 +25,17 @@ namespace
 
 } // namespace
 
-bool Selection::Contains(std::uint32_t _id) const noexcept
+bool Selection::Contains(std::uint16_t _id) const noexcept
 {
   return std::find(m_ids.begin(), m_ids.end(), _id) != m_ids.end();
 }
 
-void Selection::Set(std::span<const std::uint32_t> _ids)
+void Selection::Set(std::span<const std::uint16_t> _ids)
 {
   m_ids.assign(_ids.begin(), _ids.end());
 }
 
-void Selection::Add(std::uint32_t _id)
+void Selection::Add(std::uint16_t _id)
 {
   if (_id != INVALID_ENTITY_ID && !Contains(_id))
   {
@@ -43,7 +43,7 @@ void Selection::Add(std::uint32_t _id)
   }
 }
 
-void Selection::Toggle(std::uint32_t _id)
+void Selection::Toggle(std::uint16_t _id)
 {
   if (_id == INVALID_ENTITY_ID)
   {
@@ -63,7 +63,7 @@ void Selection::Toggle(std::uint32_t _id)
 void Selection::Retain(std::span<const SceneEntity> _entities)
 {
   std::erase_if(m_ids,
-                [_entities](std::uint32_t _id)
+                [_entities](std::uint16_t _id)
                 {
                   return std::find_if(_entities.begin(), _entities.end(),
                                       [_id](const SceneEntity& _target) { return _target.id == _id; }) == _entities.end();
@@ -116,14 +116,14 @@ void Selection::EndDrag(std::span<const SceneEntity> _entities, const PlaneMappi
     const XMFLOAT2 cornerA = PixelsToNdc(m_startX, m_startY, _viewportWidth, _viewportHeight);
     const XMFLOAT2 cornerB = PixelsToNdc(m_currentX, m_currentY, _viewportWidth, _viewportHeight);
 
-    std::vector<std::uint32_t> inside;
+    std::vector<std::uint16_t> inside;
     PickBox(_entities, _mapping, cornerA, cornerB, inside);
 
     if (!additive)
     {
       m_ids.clear();
     }
-    for (const std::uint32_t id : inside)
+    for (const std::uint16_t id : inside)
     {
       Add(id); // Add, not assign: a shift-box merges rather than replaces.
     }
@@ -134,7 +134,7 @@ void Selection::EndDrag(std::span<const SceneEntity> _entities, const PlaneMappi
   const XMFLOAT2 ndc = PixelsToNdc(m_currentX, m_currentY, _viewportWidth, _viewportHeight);
   const XMFLOAT2 planePoint{_mapping.origin.x + _mapping.rightPerNdc.x * ndc.x + _mapping.upPerNdc.x * ndc.y,
                             _mapping.origin.y + _mapping.rightPerNdc.y * ndc.x + _mapping.upPerNdc.y * ndc.y};
-  const std::uint32_t hit = PickPoint(_entities, planePoint, _minRadiusMetres);
+  const std::uint16_t hit = PickPoint(_entities, planePoint, _minRadiusMetres);
 
   if (additive)
   {
